changning ints to bigints on migration table

+++ db/migrate/20210119195246_create_locations.rb
@@ -4,15 +4,15 @@ class CreateLocations < ActiveRecord::Migration[6.0]
       t.string :country
       t.string :flag
       t.string :ISO
-      t.integer :confirmed
-      t.integer :deaths
-      t.integer :active
-      t.integer :recovered
+      t.bigint :confirmed
+      t.bigint :deaths
+      t.bigint :active
+      t.bigint :recovered
       t.float :lat
       t.float :lon
       t.string :date
-      t.integer :user_id
-      t.integer :vaccinations
+      t.bigint :user_id
+      t.bigint :vaccinations
 
       t.timestamps
     end
